refactor: remove global LenisProvider, scope Lenis to landing/about/contact pages
- Delete LenisProvider (was global, caused scroll issues in modals/chat/workspace)
- Create LenisRoot component for per-page opt-in smooth scroll
- Wrap home, about, contact pages with LenisRoot
- Remove data-lenis-prevent from UserProfileModal, WorkspaceSettingsModal, ModuleSettingsModal
