+++ src/app/(site)/about/page.jsx
@@ -1,14 +1,17 @@
 // src/app/about/page.jsx
+import LenisRoot from "@/components/LenisRoot";
 import AboutSection from "@/components/AboutUs/Hero&Features";
 import StoryAndCTA from "@/components/AboutUs/Story&CTA";
 import TeamSection from "@/components/AboutUs/TeamSection";
 
 export default function AboutPage() {
   return (
-    <main>
-      <AboutSection />
-      <TeamSection />
-      <StoryAndCTA />
-    </main>
+    <LenisRoot>
+      <main>
+        <AboutSection />
+        <TeamSection />
+        <StoryAndCTA />
+      </main>
+    </LenisRoot>
   );
 }
