+++ src/app/(site)/contact/page.jsx
@@ -1,11 +1,14 @@
+import LenisRoot from "@/components/LenisRoot";
 import Contact from "@/components/Contact/Contact";
 import FAQSection from "@/components/Contact/FAQSection";
 
 export default function Page() {
   return (
-    <main>
-      <Contact />
-      <FAQSection />
-    </main>
+    <LenisRoot>
+      <main>
+        <Contact />
+        <FAQSection />
+      </main>
+    </LenisRoot>
   );
 }
