+++ src/components/Providers.jsx
@@ -1,12 +1,7 @@
 "use client";
 
 import { AuthProvider } from "@/context/AuthProvider";
-import LenisProvider from "@/components/LenisProvider";
 
 export default function Providers({ children }) {
-  return (
-    <LenisProvider>
-      <AuthProvider>{children}</AuthProvider>
-    </LenisProvider>
-  );
+  return <AuthProvider>{children}</AuthProvider>;
 }
